Update to implement role-based accees control,and playlist managment

+++ src/main/java/org/media_player/application/services/MediaFileService.java
@@ -1,23 +1,28 @@
 package org.media_player.application.services;
 
-import org.media_player.application.exceptions.MediaFileNotFoundException;
+import org.media_player.application.exceptions.MediaFileException;
 import org.media_player.domain.abstractions.MediaFileRepository;
 import org.media_player.domain.entities.media.MediaFile;
+import org.media_player.domain.entities.user.User;
 import org.media_player.domain.factories.MediaFileFactoryImpl;
 
 import java.util.List;
-import java.util.Optional;
 
 public class MediaFileService {
     private final MediaFileRepository mediaFileRepository;
     private final MediaFileFactoryImpl mediaFileFactoryImpl;
+    private final AuthorizationService authorizationService;
 
-    public MediaFileService(MediaFileRepository mediaFileRepository, MediaFileFactoryImpl mediaFileFactoryImpl) {
+    public MediaFileService(MediaFileRepository mediaFileRepository, MediaFileFactoryImpl mediaFileFactoryImpl, AuthorizationService authorizationService) {
         this.mediaFileRepository = mediaFileRepository;
         this.mediaFileFactoryImpl = mediaFileFactoryImpl;
+        this.authorizationService = authorizationService;
     }
 
-    public void saveMediaFile(String mediaType, String fileName, String filePath, String fileExtension) {
+    public void saveMediaFile(User user, String mediaType, String fileName, String filePath, String fileExtension) {
+        if(!authorizationService.isAdmin(user)){
+            throw new MediaFileException("User is not authorized to save media file");
+        }
         MediaFile mediaFile = mediaFileFactoryImpl.createMediaFile(mediaType, fileName, filePath, fileExtension);
         mediaFileRepository.saveMediaFile(mediaFile);
 
@@ -27,11 +32,14 @@ public class MediaFileService {
         if (mediaFileRepository.getMediaFile(fileName).isPresent()) {
             return mediaFileRepository.getMediaFile(fileName).get();
         } else {
-            throw new MediaFileNotFoundException("Media file not found with name: " + fileName);
+            throw new MediaFileException("Media file not found with name: " + fileName);
         }
     }
 
-    public void deleteMediaFile(String fileName) {
+    public void deleteMediaFile(User user, String fileName) {
+        if(!authorizationService.isAdmin(user)){
+            throw new MediaFileException("User is not authorized to delete media file");
+        }
         mediaFileRepository.deleteMediaFile(fileName);
     }
 
@@ -39,7 +47,10 @@ public class MediaFileService {
         return mediaFileRepository.getAllMediaFiles();
     }
 
-    public void updateMediaFile(String mediaType, String fileName, String filePath, String fileExtension) {
+    public void updateMediaFile(User user,String mediaType, String fileName, String filePath, String fileExtension) {
+        if(!authorizationService.isAdmin(user)){
+            throw new MediaFileException("User is not authorized to update media file");
+        }
         MediaFile mediaFile = mediaFileFactoryImpl.createMediaFile(mediaType, fileName, filePath, fileExtension);
         mediaFileRepository.updateMediaFile(mediaFile);
     }
